requirements.txt
accelerate==1.0.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
diffusers==0.11.1
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
ftfy==6.3.0
huggingface-hub==0.24.7
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pillow==10.2.0
psutil==6.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
six==1.16.0
sympy==1.13.1
tokenizers==0.20.1
torch==2.5.0+cu121
torchaudio==2.5.0+cu121
torchvision==0.20.0+cu121
tqdm==4.66.5
transformers @ git+https://github.com/huggingface/transformers.git@940a6bd343cfd2ff4f4425b4cbc548d1e1d316da
triton==3.1.0
typing_extensions==4.9.0
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.20.2
